warning fixes for viewer directory
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23638 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/viewer/main/ViewActions.C
@@ -325,7 +325,7 @@ void
 SaveViewAction::DeleteViews()
 {
     // Delete the views
-    for(int i = 0; i < views.size(); ++i)
+    for(size_t i = 0; i < views.size(); ++i)
     {
         if(views[i].viewType == VIEWCurve)
         {
@@ -457,7 +457,7 @@ SaveViewAction::CopyFrom(const ViewerActionBase *obj)
         // Delete the views from the interface.
         DeleteViewsFromInterface();
 
-        for(int i = 0; i < saveView->views.size(); ++i)
+        for(size_t i = 0; i < saveView->views.size(); ++i)
         {
             void *newView = 0;
 
@@ -515,7 +515,7 @@ SaveViewAction::CopyFrom(const ViewerActionBase *obj)
 void
 SaveViewAction::SaveCurrentView()
 {
-    if(views.size() < MAX_SAVED_VIEWS)
+    if(views.size() < (size_t)MAX_SAVED_VIEWS)
     {
         void *saveView;
         int vt;
@@ -649,7 +649,7 @@ SaveViewAction::AddNewView(void *v, int vt)
 void
 SaveViewAction::UseSavedView(int index)
 {
-    if(index >= 0 && index < views.size())
+    if(index >= 0 && (size_t)index < views.size())
     {
         if(views[index].viewType == VIEWCurve)
         {
@@ -797,7 +797,7 @@ SaveViewAction::CreateNode(DataNode *parentNode)
 
     // Add a node for each view that is not the default view.
     intVector viewTypes;
-    for(int i = 0; i < views.size(); ++i)
+    for(size_t i = 0; i < views.size(); ++i)
     {
         if(views[i].viewType == VIEWCurve)
         {
